Fix payment being updated after completed.
See: https://github.com/boomerdigital/solidus_avatax_certified/issues/51

Updating the payment here will cause issues with multiple payments and
is redundant.

Since the payment has already been authorized and captured, there
is no logical reason to update the payment amount to be other than
what the payment currently is.
