+++ circle.yml
@@ -11,7 +11,7 @@ database:
 test:
   pre:
     - gem install bundler --pre
-    - bundle exec rake test_app:
+    - echo "Y" | bundle exec rake test_app:
         environment:
           RAILS_ENV: test
           RACK_ENV: test
